refactor: split cli progress tests by class
