+++ README.md
@@ -101,6 +101,10 @@ dotnet user-secrets set "arius:<account>:key" "<key>"
 Most test projects can be run directly with `dotnet test --project <path-to-csproj>`.
 `src/Arius.E2E.Tests` also requires `ARIUS_E2E_ACCOUNT` and `ARIUS_E2E_KEY` to be set; otherwise the suite fails immediately with a configuration error.
 
+The bundled `ast-grep` skill in `.agents/skills/ast-grep/` is maintained as C#-first guidance. Its examples assume `--lang csharp` and prefer `pattern.context` plus `selector` when matching declarations such as methods or properties.
+
+The test suites are being mechanically refactored to mirror the production structure they exercise. For example, `src/Arius.Cli/Commands/Archive/` maps to `src/Arius.Cli.Tests/Commands/Archive/`, while tests for top-level CLI files such as `DisplayHelpers.cs` stay at the test-project root.
+
 ## Updating
 
 Run:
